tests/wpt/web-platform-tests/mediacapture-streams/obtaining-local-multimedia-content/navigatorusermedia/api-present.html

<!doctype html>
<html>
<head>
<title>getUserMedia: test that getUserMedia is present (with or without vendor prefix)</title>
<link rel="author" title="Dominique Hazael-Massieux" href="mailto:dom@w3.org"/>
<link rel="help" href="http://dev.w3.org/2011/webrtc/editor/getusermedia.html#navigatorusermedia">
<meta name='assert' content='Check that the getUserMedia() method is present.'/>
<meta name='flags' content='vendor-prefix, dom'/>
<link rel='stylesheet' href='/resources/testharness.css' media='all'/>
</head>
<body>
<h1 class="instructions">Description</h1>
<p class="instructions">This test checks for the presence of the
<code>navigator.getUserMedia</code> method, taking vendor prefixes into account.</p>
<div id='log'></div>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script src="/common/vendor-prefix.js" data-prefixed-objects='[{"ancestors":["navigator"], "name":"getUserMedia"}]'></script>
<script>
test(function () {
          assert_true(undefined !== navigator.getUserMedia, "navigator.getUserMedia exists");
      }, "getUserMedia() is present on navigator");
</script>
</body>
</html>
